Rename ui to legacy-ui
