+++ services/frontend-service/src/legacy-ui/ActionDetails.test.tsx
