+++ services/frontend-service/src/legacy-ui/ActionDetails.tsx
