+++ services/frontend-service/src/legacy-ui/AppBar/LocksDrawer.test.tsx
