+++ services/frontend-service/src/legacy-ui/AppBar/LocksDrawer.tsx
