+++ services/frontend-service/src/v2-ui/Routes.tsx
@@ -14,7 +14,7 @@ You should have received a copy of the GNU General Public License
 along with kuberpult.  If not, see <http://www.gnu.org/licenses/>.
 
 Copyright 2021 freiheit.com*/
-import { App as LegacyApp } from '../ui/App';
+import { App as LegacyApp } from '../legacy-ui/App';
 import { App } from './App';
 import { Routes as ReactRoutes, Route } from 'react-router-dom';
